+++ propack-bukkit/src/main/java/me/nelonn/propack/bukkit/ProPack.java
@@ -23,7 +23,7 @@ import me.nelonn.flint.path.Identifier;
 import me.nelonn.propack.bukkit.dispatcher.Dispatcher;
 import me.nelonn.propack.core.ProPackCore;
 import me.nelonn.propack.definition.Item;
-import me.nelonn.propack.bukkit.resourcepack.PackContainer;
+import me.nelonn.propack.bukkit.definition.PackContainer;
 import org.bukkit.Material;
 import org.jetbrains.annotations.NotNull;
