+++ propack-bukkit/src/main/java/me/nelonn/propack/bukkit/command/BuildCommand.java
@@ -22,8 +22,8 @@ import me.nelonn.propack.ResourcePack;
 import me.nelonn.propack.bukkit.ProPack;
 import me.nelonn.propack.bukkit.ProPackPlugin;
 import me.nelonn.propack.bukkit.Util;
-import me.nelonn.propack.bukkit.resourcepack.PackDefinition;
-import me.nelonn.propack.bukkit.resourcepack.ProjectDefinition;
+import me.nelonn.propack.bukkit.definition.PackDefinition;
+import me.nelonn.propack.bukkit.definition.ProjectDefinition;
 import me.nelonn.propack.core.builder.InternalProject;
 import org.bukkit.Bukkit;
 import org.bukkit.command.CommandSender;
