+++ propack-bukkit/src/main/java/me/nelonn/propack/bukkit/definition/PackContainer.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package me.nelonn.propack.bukkit.resourcepack;
+package me.nelonn.propack.bukkit.definition;
 
 import com.google.gson.JsonObject;
 import me.nelonn.propack.core.ProPackCore;
