+++ propack-bukkit/src/main/java/me/nelonn/propack/bukkit/definition/ProjectDefinition.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package me.nelonn.propack.bukkit.resourcepack;
+package me.nelonn.propack.bukkit.definition;
 
 import me.nelonn.propack.ResourcePack;
 import me.nelonn.propack.builder.Project;
